Move get_system_info and FtdOperations to module_utils
